refactor: add AdScripts component for managing advertisement scripts in the application
